+++ app/src/main/java/com/make_your_skill/ui/components/bars/BottonBarNav.kt
@@ -1,17 +1,15 @@
-package com.make_your_skill.ui.components
+package com.make_your_skill.ui.components.bars
 
-import android.graphics.Color
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavHostController
 import com.make_your_skill.R
+import com.make_your_skill.ui.components.buttons.CustomIconButton
 import com.make_your_skill.ui.navigation.AppRoutes
 import com.make_your_skill.ui.theme.Purple
