+++ app/src/main/java/com/make_your_skill/ui/components/buttons/CustomTextField.kt
@@ -1,4 +1,4 @@
-package com.make_your_skill.ui.components
+package com.make_your_skill.ui.components.buttons
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
