+++ app/src/main/java/com/make_your_skill/ui/components/popUps/AddSkillPopUp.kt
@@ -1,4 +1,4 @@
-package com.make_your_skill.ui.components
+package com.make_your_skill.ui.components.popUps
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.material3.AlertDialog
@@ -7,12 +7,17 @@ import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import com.make_your_skill.dataClasses.skills.skillDataClass
+import com.make_your_skill.ui.components.buttons.CustomTextField
+import com.make_your_skill.ui.components.dropDowns.skillsDropDown
 
 @Composable
-fun addInterestsPopUp(
+fun addSkillPopUp(
     onDismissRequest: () -> Unit,//Cerramos popup
     onConfirmation: () -> Unit,//Confirmamos el popup
     dialogTitle: String,//Titulo del popup
+    onPriceAddChange: (String) -> Unit,//Cuando cambio el input del precio
+    priceLabel: String,//Label del precio
+    priceText: String,
     skillsList: List<skillDataClass>,
     onSkillAddChange: (Int) -> Unit
 ) {
@@ -25,6 +30,7 @@ fun addInterestsPopUp(
               modifier = Modifier
           ) {
               skillsDropDown(skillsList, onSkillAddChange)
+              CustomTextField(priceText,onPriceAddChange,priceLabel, onSubmit = {}, false)
           }
         },
         onDismissRequest = {
