+++ app/src/main/java/com/make_your_skill/ui/components/text/textFileds/ClickeableText.kt
@@ -1,12 +1,10 @@
-package com.make_your_skill.ui.components
+package com.make_your_skill.ui.components.text.textFileds
 
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 
 @Composable
