+++ app/src/main/java/com/make_your_skill/ui/screens/birthday/BirthdayScreen.kt
@@ -24,10 +24,10 @@ import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavHostController
 import com.make_your_skill.helpers.cookies.InMemoryCookieJar
 import com.make_your_skill.helpers.functions.getNextDate
-import com.make_your_skill.ui.components.CustomButton
+import com.make_your_skill.ui.components.buttons.CustomButton
 import com.make_your_skill.ui.components.DatePickerDocked
 import com.make_your_skill.ui.components.convertMillisToDate
-import com.make_your_skill.ui.components.ScreenTitleText
+import com.make_your_skill.ui.components.text.ScreenTitleText
 import com.make_your_skill.ui.navigation.AppRoutes
 import com.make_your_skill.ui.screens.createNewAccount.CreateNewAcoountViewModel
 import com.make_your_skill.ui.screens.singIn.SingInViewModel
