+++ app/src/main/java/com/make_your_skill/ui/screens/createNewAccount/CreateNewAccountScreen.kt
@@ -11,11 +11,8 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
@@ -25,13 +22,10 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavHostController
 import com.make_your_skill.R
-import com.make_your_skill.ui.components.CustomButton
-import com.make_your_skill.ui.components.TextInputLogin
-import com.make_your_skill.ui.navigation.AppRoutes
-import com.make_your_skill.ui.screens.singIn.SingInViewModel
+import com.make_your_skill.ui.components.buttons.CustomButton
+import com.make_your_skill.ui.components.text.textFileds.TextInputLogin
 import com.make_your_skill.ui.theme.BackgroundColor2
