+++ app/src/main/java/com/make_your_skill/ui/screens/firstName/FirstNameScreen.kt
@@ -12,20 +12,16 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
-import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavHostController
-import com.make_your_skill.ui.components.CustomButton
-import com.make_your_skill.ui.components.CustomTextField
-import com.make_your_skill.ui.components.ScreenTitleText
+import com.make_your_skill.ui.components.buttons.CustomButton
+import com.make_your_skill.ui.components.buttons.CustomTextField
+import com.make_your_skill.ui.components.text.ScreenTitleText
 import com.make_your_skill.ui.screens.createNewAccount.CreateNewAcoountViewModel
 
 @Composable
