+++ app/src/main/java/com/make_your_skill/ui/screens/matchSearch/MatchSearchScreen.kt
@@ -16,12 +16,13 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.platform.LocalConfiguration
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.rememberNavController
-import com.make_your_skill.ui.components.*
+import com.make_your_skill.ui.components.buttons.CustomButton
+import com.make_your_skill.ui.components.buttons.CustomTextField
+import com.make_your_skill.ui.components.sliders.RangeSlider
 import com.make_your_skill.ui.navigation.AppRoutes
 import com.make_your_skill.ui.theme.*
