+++ app/src/main/java/com/make_your_skill/ui/screens/profileSettings/ProfileSettingsScreen.kt
@@ -7,7 +7,6 @@ import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -22,7 +21,7 @@ import androidx.compose.ui.unit.sp
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.rememberNavController
 import com.make_your_skill.R
-import com.make_your_skill.ui.components.ClickableText
+import com.make_your_skill.ui.components.text.textFileds.ClickableText
 import com.make_your_skill.ui.theme.BackgroundColor2
 import com.make_your_skill.ui.theme.styleTitle
