+++ app/src/main/java/com/make_your_skill/ui/screens/skill/SkillScreen.kt
@@ -22,15 +22,13 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavHostController
-import com.make_your_skill.dataClasses.skills.skillDataClass
 import com.make_your_skill.helpers.cookies.InMemoryCookieJar
-import com.make_your_skill.ui.components.CustomButton
-import com.make_your_skill.ui.components.addSkillPopUp
-import com.make_your_skill.ui.components.ScreenTitleText
+import com.make_your_skill.ui.components.buttons.CustomButton
+import com.make_your_skill.ui.components.popUps.addSkillPopUp
+import com.make_your_skill.ui.components.text.ScreenTitleText
 import com.make_your_skill.ui.components.skill
 import com.make_your_skill.ui.screens.singIn.SingInViewModel
 import com.make_your_skill.ui.theme.DarkPurple
-import kotlinx.coroutines.delay
 
 
 @Composable
